Move static_vm_spec to its own file
PiperOrigin-RevId: 879840203
